requirements.txt
anyio==4.0.0
astor==0.8.1
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
decorator==5.1.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
imageio==2.32.0
iniconfig==2.0.0
lazy_loader==0.3
lmdb==1.4.1
more-itertools==10.1.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
Pillow==10.0.1
pluggy==0.13.1
protobuf==4.25.0
py==1.11.0
pyclipper==1.3.0.post5
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.1.3
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.5.2
rarfile==4.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.3
selenium==4.15.2
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
termcolor==2.3.0
tifffile==2023.9.26
tomli==2.0.1
tqdm==4.66.1
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.9
wsproto==1.2.0
